# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Events API - Luma-Inspired Event Management",
        "description": "Open-source headless Laravel event management API from Maak. Features organizations, events, ticket types, registrations, QR check-ins, waitlist management, and team member invitations.",
        "contact": {
            "name": "Maak",
            "url": "https://github.com/maakwizardry/events"
        },
        "version": "1.1.0"
    },
    "servers": [
        {
            "url": "http://events.test/api/v1",
            "description": "Local Development Server"
        },
        {
            "url": "http://localhost:8000/api/v1",
            "description": "PHP Built-in Server"
        }
    ],
    "paths": {
...
        "/events/{event}/check-in": {
            "post": {
                "tags": [
                    "Check-In"
                ],
                "summary": "Check in attendee",
                "description": "Check in an attendee using their QR code. Requires check-in permissions for the event.",
                "operationId": "06c0897ab8b7588ba76e3c1df89497e0",
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "description": "Event UUID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "qr_code_data"
                                ],
                                "properties": {
                                    "qr_code_data": {
                                        "type": "string",
                                        "example": "QR_CODE_STRING"
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "Main Entrance",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Check-in successful",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Check-in successful"
                                        },
                                        "registration": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Not authorized to check in attendees"
                    },
                    "404": {
                        "description": "Invalid QR code"
                    },
                    "422": {
                        "description": "Already checked in or invalid registration status"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/events/{event}/check-in/stats": {
            "get": {
                "tags": [
                    "Check-In"
                ],
                "summary": "Get check-in statistics",
                "description": "Get comprehensive check-in statistics and analytics for an event. Requires check-in permissions.",
                "operationId": "7290e233b9408772a4a23c61b905234e",
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "description": "Event UUID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Check-in statistics",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "event": {
                                            "properties": {
                                                "uuid": {
                                                    "type": "string"
                                                },
                                                "name": {
                                                    "type": "string"
                                                },
                                                "starts_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "summary": {
                                            "properties": {
                                                "total_registrations": {
                                                    "type": "integer",
                                                    "example": 100
                                                },
                                                "checked_in": {
                                                    "type": "integer",
                                                    "example": 75
                                                },
                                                "pending_check_in": {
                                                    "type": "integer",
                                                    "example": 25
                                                },
                                                "waitlisted": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "cancelled": {
                                                    "type": "integer",
                                                    "example": 5
                                                },
                                                "check_in_rate": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": 75
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "check_ins_by_hour": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        },
                                        "ticket_type_breakdown": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Not authorized to view check-in statistics"
                    },
                    "404": {
                        "description": "Event not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/events/{event}/registrations": {
            "get": {
                "tags": [
                    "Check-In"
                ],
                "summary": "List event registrations",
                "description": "Get all registrations for an event with filtering and search. Requires registration management permissions.",
                "operationId": "a0857f357e038e822827b14ba87ba02c",
                "parameters": [
                    {
                        "name": "event",
                        "in": "path",
                        "description": "Event UUID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by registration status",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "confirmed",
                            "enum": [
                                "confirmed",
                                "waitlisted",
                                "cancelled"
                            ]
                        }
                    },
                    {
                        "name": "checked_in",
                        "in": "query",
                        "description": "Filter by check-in status",
                        "required": false,
                        "schema": {
                            "type": "boolean",
                            "example": true
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search by attendee name or email",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "john"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number for pagination",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of registrations",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Not authorized to manage registrations"
                    },
                    "404": {
                        "description": "Event not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Use the token from POST /auth/login or POST /auth/register",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}